# docs/swagger.yaml
basePath: /dev-api
definitions:
  auth.LoginRequest:
    properties:
      bizType:
        allOf:
        - $ref: '#/definitions/captcha.BizTypeCaptcha'
        description: 验证码类型
      code:
        description: 验证码
        type: string
      id:
        description: 验证码
        type: string
      password:
        description: 密码
        maxLength: 20
        minLength: 3
        type: string
      username:
        description: 用户账号
        maxLength: 50
        minLength: 3
        type: string
    required:
    - bizType
    - code
    - id
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      token:
        description: 登录令牌
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      password:
        description: 密码
        maxLength: 20
        minLength: 3
        type: string
      username:
        description: 用户账号
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  captcha.BizTypeCaptcha:
    enum:
    - BizCaptchaLogin
    type: string
    x-enum-comments:
      BizTypeCaptchaLogin: 验证码密码登录
    x-enum-varnames:
    - BizTypeCaptchaLogin
  captcha.TypeCaptcha:
    enum:
    - 1
    type: integer
    x-enum-comments:
      DigitIotaCaptcha: 数字字母验证码
    x-enum-varnames:
    - DigitIotaCaptcha
  dict.TypeConst:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      TypeConst0: 普通字典
      TypeConst1: 系统字典
      TypeConst2: 枚举字典
    x-enum-varnames:
    - TypeConst0
    - TypeConst1
    - TypeConst2
  dict.TypeValueConst:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      TypeValueConst0: 字符串
      TypeValueConst1: 整型
      TypeValueConst2: 布尔
    x-enum-varnames:
    - TypeValueConst0
    - TypeValueConst1
    - TypeValueConst2
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
      success:
        description: 是否成功
        type: boolean
    type: object
  third.CaptchaResponse:
    properties:
      code:
        description: 验证码
        type: string
      id:
        description: 验证码Id
        type: string
      img:
        description: 验证码图片
        type: string
    type: object
  tools.Request:
    properties:
      code:
        description: 字典编码
        maxLength: 100
        type: string
      name:
        description: 字典名称
        maxLength: 100
        type: string
      remark:
        description: 备注
        maxLength: 255
        type: string
      type:
        allOf:
        - $ref: '#/definitions/dict.TypeConst'
        description: 字典类型
      typeValue:
        allOf:
        - $ref: '#/definitions/dict.TypeValueConst'
        description: 字典类型值
      version:
        description: 版本
        type: integer
    required:
    - code
    - name
    type: object
host: localhost:8080
info:
  contact:
    email: 2224693191@qq.com
    name: CJiaの用心
    url: http://www.swagger.io/support
  description: CarefulAdmin在线接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: CarefulAdmin
  version: "1.0"
paths:
  /v1/auth/password-login:
    post:
      consumes:
      - application/json
      description: 图形验证码密码登录
      parameters:
      - description: 参数
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 图形验证码密码登录
      tags:
      - 认证管理
  /v1/auth/password-register:
    post:
      consumes:
      - application/json
      description: 账号密码注册
      parameters:
      - description: 参数
        in: body
        name: RegisterRequest
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 账号密码注册
      tags:
      - 认证管理
  /v1/third/generateCaptcha:
    get:
      consumes:
      - application/json
      description: 生成指定业务验证码
      parameters:
      - description: 业务类型
        enum:
        - BizCaptchaLogin
        in: query
        name: bizType
        required: true
        type: string
        x-enum-comments:
          BizTypeCaptchaLogin: 验证码密码登录
        x-enum-varnames:
        - BizTypeCaptchaLogin
      - description: 验证码类型
        enum:
        - 1
        in: query
        name: type
        required: true
        type: integer
        x-enum-comments:
          DigitIotaCaptcha: 数字字母验证码
        x-enum-varnames:
        - DigitIotaCaptcha
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/third.CaptchaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 生成指定业务验证码
      tags:
      - 第三方业务管理
  /v1/tools/dict/batchDelete:
    post:
      consumes:
      - application/json
      description: 批量删除字典
      parameters:
      - description: ID数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 批量删除字典
      tags:
      - 字典管理
  /v1/tools/dict/create:
    post:
      consumes:
      - application/json
      description: 创建字典
      parameters:
      - description: 字典信息
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/tools.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 创建字典
      tags:
      - 字典管理
  /v1/tools/dict/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除字典
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 删除字典
      tags:
      - 字典管理
  /v1/tools/dict/getById/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取字典
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 根据ID获取字典
      tags:
      - 字典管理
  /v1/tools/dict/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入字典
      parameters:
      - description: 文件(支持xlsx/csv格式)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 导入字典
      tags:
      - 字典管理
  /v1/tools/dict/listAll:
    get:
      consumes:
      - application/json
      description: 获取所有字典
      parameters:
      - description: 创建人
        in: query
        name: creator
        type: string
      - description: 修改人
        in: query
        name: modifier
        type: string
      - default: true
        description: 状态
        in: query
        name: status
        type: boolean
      - description: 字典名称
        in: query
        name: name
        type: string
      - description: 字典编码
        in: query
        name: code
        type: string
      - default: -1
        description: 字典类型
        in: query
        name: type
        type: integer
      - default: -1
        description: 字典数据类型
        in: query
        name: typeValue
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 获取所有字典
      tags:
      - 字典管理
  /v1/tools/dict/listPage:
    get:
      consumes:
      - application/json
      description: 分页获取字典
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 创建人
        in: query
        name: creator
        type: string
      - description: 修改人
        in: query
        name: modifier
        type: string
      - default: true
        description: 状态
        in: query
        name: status
        type: boolean
      - description: 字典名称
        in: query
        name: name
        type: string
      - description: 字典编码
        in: query
        name: code
        type: string
      - default: -1
        description: 字典类型
        in: query
        name: type
        type: integer
      - default: -1
        description: 字典数据类型
        in: query
        name: typeValue
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 分页获取字典
      tags:
      - 字典管理
  /v1/tools/dict/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新字典
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 字典信息
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/tools.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - LoginToken: []
      summary: 更新字典
      tags:
      - 字典管理
securityDefinitions:
  LoginToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"